chnages

+++ routes/bookmarks.js
@@ -3,49 +3,48 @@ const router = express.Router();
 const BookmarksSchema = require("../models/bookmarks_schema");
 const jwt = require("jsonwebtoken");
 require("dotenv").config();
-
+const CheckAuth = require("./../functions/check_auth");
 
 //Get bookmarks
 router.get("/", async (req, res) => {
+  const check = await CheckAuth(req, res);
 
-  const token = req.cookies.auth_token || req.body.token || req.headers["x-auth-token"];
-
-  if (token === undefined || token === null || token === "") {
-    return res.status(401).json({ message: "Unauthorized" });
+  if (check.auth === false) {
+    return res.status(401).json({ message: "Unauthorized", auth: false });
   }
-
-  const decoded = jwt.verify(token, process.env.JWT_SECRET);
 
   try {
-    const bookmarks = await BookmarksSchema.find({ user_id: decoded.id });
+    const bookmarks = await BookmarksSchema.find({ user_id: check.data._id }).populate("news_id");
     res.status(200).json(bookmarks);
-  }
-  catch (error) {
+  } catch (error) {
     res.status(500).json({ message: error.message });
   }
 });
 
 //Add bookmark
-router.post("/", (req, res) => {
-  const token = req.cookies.auth_token || req.body.token || req.headers["x-auth-token"];
+router.post("/", async (req, res) => {
+  const check = await CheckAuth(req, res);
 
-  if (token === undefined || token === null || token === "") {
-    return res.status(401).json({ message: "Unauthorized" });
+  if (check.auth === false) {
+    return res.status(401).json({ message: "Unauthorized", auth: false });
   }
 
-  const decoded = jwt.verify(token, process.env.JWT_SECRET);
+  //Check if bookmark already exists
+  const bookmarks = await BookmarksSchema.findOne({
+    user_id: check.data._id,
+    news_id: req.body.news_id,
+  }).exec();
 
   //Check if bookmark already exists
-  const bootkmarks = BookmarksSchema.findOne({ user_id: decoded.id, news_id: req.body.news_id });
-
-  if (bootkmarks) {
-    return res.status(400).json({ message: "Bookmark already exists" });
+  if (bookmarks) {
+    return res.status(409).json({ message: "Bookmark already exists" });
   }
 
   const bookmark = new BookmarksSchema({
-    user_id: decoded.id,
+    user_id: check.data._id,
     news_id: req.body.news_id,
   });
+
   bookmark
     .save()
     .then((result) => {
@@ -55,14 +54,40 @@ router.post("/", (req, res) => {
       });
     })
     .catch((err) => {
+      let message = "";
+      switch (err.code) {
+        case 11000:
+          message = "Bookmark already exists";
+          break;
+        default:
+          message = "Error saving bookmark";
+      }
       res.status(500).json({
-        message: err.message,
+        message: message,
       });
     });
 });
 
 //Delete bookmark
-router.delete("/:id", (req, res) => {
+router.delete("/:id", async (req, res) => {
+  const check = await CheckAuth(req, res);
+
+  if (check.auth === false) {
+    return res.status(401).json({ message: "Unauthorized", auth: false });
+  }
+
+  const bookmarks = await BookmarksSchema.findOne({
+    _id: req.params.id,
+  }).exec();
+
+  if (!bookmarks) {
+    return res.status(404).json({ message: "Bookmark not found" });
+  }
+
+  if (bookmarks.user_id != check.data._id) {
+    return res.status(403).json({ message: "Not created by you" });
+  }
+
   BookmarksSchema.deleteOne({ _id: req.params.id })
     .then((result) => {
       res.status(200).json({
